added admin route and showed sidebar items on userbased

+++ src/Pages/Dashboard/MyProducts/MyProducts.jsx
@@ -26,7 +26,9 @@ const MyProducts = () => {
         return <Spinner></Spinner>;
       }
 
-      const handleDelete = (id) => {
+  const handleDelete = (id) => {
+         const agree = window.confirm("Are you sure to Delete this product?");
+    if (agree) {
         fetch(`http://localhost:5000/my-products/${id}`, {
           method: "DELETE",
         })
@@ -41,11 +43,14 @@ const MyProducts = () => {
             console.log(error);
             setLoading(false);
           });
+     }
       };
   const handleAdvertise = (id) => {
+     const agree = window.confirm("Are you sure to advertise this product?");
+    if (agree) {
     const choosedProduct = products.find((product) => product._id === id);
-  
-    console.log(choosedProduct);
+      // choosedProduct.status = 'advertise';
+    // console.log(choosedProduct);
         fetch(`http://localhost:5000/advertise`, {
           method: "POST",
           headers: {
@@ -57,7 +62,7 @@ const MyProducts = () => {
           .then((data) => {
             if (data.acknowledged) {
               toast.success("Product added to advertised successfuly");
-              
+              setRefresh(!refresh);
             } else{
               toast.success("Product is already advertised");
             }
@@ -66,8 +71,9 @@ const MyProducts = () => {
             console.log(error);
            
           });
+           }
       };
-    // console.log(products);
+    console.log(products);
     return (
       <div className="container px-4 sm:px-8 max-w-3xl flex justify-left items-base mt-10 content-left overflow-x-auto">
         <div className="py-8">
@@ -86,14 +92,26 @@ const MyProducts = () => {
                       scope="col"
                       className="px-5 py-3 bg-white  border-b border-gray-200 text-gray-800  text-left text-sm uppercase font-normal"
                     >
-                      Role
+                      Name
                     </th>
 
                     <th
                       scope="col"
                       className="px-5 py-3 bg-white  border-b border-gray-200 text-gray-800  text-left text-sm uppercase font-normal"
                     >
-                      status
+                      Price
+                    </th>
+                    <th
+                      scope="col"
+                      className="px-5 py-3 bg-white  border-b border-gray-200 text-gray-800  text-left text-sm uppercase font-normal"
+                    >
+                      advertise
+                    </th>
+                    <th
+                      scope="col"
+                      className="px-5 py-3 bg-white  border-b border-gray-200 text-gray-800  text-left text-sm uppercase font-normal"
+                    >
+                      Status
                     </th>
                     <th
                       scope="col"
@@ -117,29 +135,62 @@ const MyProducts = () => {
                               />
                             </Link>
                           </div>
-                          <div className="ml-3">
-                            <p className="text-gray-900 whitespace-no-wrap">
-                              {product.name}
-                            </p>
-                          </div>
+                        </div>
+                      </td>
+                      <td>
+                        <div className="ml-3">
+                          <p className="text-gray-900 whitespace-no-wrap">
+                            {product.name}
+                          </p>
+                        </div>
+                      </td>
+                      <td>
+                        <div className="ml-3">
+                          <p className="text-gray-900 whitespace-no-wrap">
+                            {product.sellPrice} tk
+                          </p>
                         </div>
                       </td>
                       <td className="px-5 py-5 border-b border-gray-200 bg-white text-sm">
                         <p
-                          className="text-gray-900 whitespace-no-wrap bg-green-300 px-1 rounded-full"
-                          onClick={() => handleAdvertise(product._id)}
+                          className={`text-gray-900 whitespace-no-wrap ${
+                            product.status === "paid" && "bg-gray-300 "
+                          } ${
+                            product.status === "advertised"
+                              ? "bg-gray-300 "
+                              : "bg-orange-300"
+                          } px-1 rounded-full cursor-pointer`}
+                          {...(product.status === "paid"
+                            ? { disabled: true }
+                            : { onClick: () => handleAdvertise(product._id) })}
                         >
-                          Advertise
+                          {product.status === "advertised"
+                            ? "Advertised"
+                            : "advertise"}
                         </p>
                       </td>
 
                       <td className="px-5 py-5 border-b border-gray-200 bg-white text-sm">
                         <span className="relative inline-block px-3 py-1 font-semibold text-green-900 leading-tight">
-                          <span
-                            aria-hidden="true"
-                            className="absolute inset-0 bg-orange-200 opacity-50 rounded-full"
-                          ></span>
-                          <span className="relative">Unsold</span>
+                          {product.status === "paid" ? (
+                            <div>
+                              <span
+                                aria-hidden="true"
+                                className="absolute inset-0 bg-green-900 opacity-50 rounded-full"
+                              ></span>
+                              <span className="relative text-white text-lg">
+                                Sold{" "}
+                              </span>
+                            </div>
+                          ) : (
+                            <div>
+                              <span
+                                aria-hidden="true"
+                                className="absolute inset-0 bg-orange-200 opacity-50 rounded-full"
+                              ></span>
+                              <span className="relative">Available </span>
+                            </div>
+                          )}
                         </span>
                       </td>
                       <td className="px-5 py-5 border-b border-gray-200 bg-white text-sm">
